Replace dnstats.net with dark.fail in tools.md
dnstats.net seems to be different than how it used to be. The links within it are likely pointed to clone markets with the purpose of scamming users that are trying to find specific links. dark.fail is closest to what dnstats was in terms of providing up to date verified content and uptime monitors for each link

What dnstats used to be https://web.archive.org/web/20191010053316/https://dnstats.net/